# Fix SDPA context manager perf regression