@mobicham 2024-10-25
Thank you Mark, works great!